Cleanup, renaming and javadoc fixes
