+++ zoltar-core/src/test/java/com/spotify/zoltar/loaders/ModelMemoizerTest.java
@@ -31,7 +31,7 @@ import java.util.concurrent.TimeoutException;
 import java.util.concurrent.atomic.AtomicInteger;
 import org.junit.Test;
 
-public class MemoizerTest {
+public class ModelMemoizerTest {
 
   static class DummyModel implements Model<Object> {
     private final AtomicInteger inc;
@@ -59,7 +59,7 @@ public class MemoizerTest {
   @Test
   public void memoize() throws InterruptedException, ExecutionException, TimeoutException {
     final ModelLoader<DummyModel> loader =
-        ModelLoader.lift(DummyModel::new).with(Memoizer::memoize);
+        ModelLoader.lift(DummyModel::new).with(ModelMemoizer::memoize);
 
     final Duration duration = Duration.ofMillis(1000);
     loader.get(duration).instance();
